Next SVN commits

+++ src/hardware/vga_crtc.cpp
@@ -275,7 +275,7 @@ void vga_write_p3d5(Bitu port,Bitu val,Bitu iolen) {
 		break;
 	case 0x12:	/* Vertical Display End Register */
 		if (val!=crtc(vertical_display_end)) {
-			if (abs((Bits)val-(Bits)crtc(vertical_display_end))<3) {
+			if (abs(static_cast<int>((Bits)val-(Bits)crtc(vertical_display_end)))<3) {
 				// delay small vde changes a bit to avoid screen resizing
 				// if they are reverted in a short timeframe
 				PIC_RemoveEvents(VGA_SetupDrawing);
